refactor(test): split promotion e2e helpers
